% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_univariate_stratified.R
\name{get_mh}
\alias{get_mh}
\title{These are internal functions written by Zhian Kamvar.
This was adapted from the code in the epiR::epi.2by2 version 1.0-2
Many of the changes involve abstracting repetative routines into functions,
splitting nested calculations into separate variables, and modifying
variable names.}
\usage{
get_mh(
  arr,
  measure = "OR",
  conf = 0.95,
  exposurelength = NULL,
  stratalength = NULL
)
}
\description{
These are internal functions written by Zhian Kamvar.
This was adapted from the code in the epiR::epi.2by2 version 1.0-2
Many of the changes involve abstracting repetative routines into functions,
splitting nested calculations into separate variables, and modifying
variable names.
}
