% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_cfr.R
\name{gt_remove_stat}
\alias{gt_remove_stat}
\title{A gtsummary wrapper function that takes a gtsummary object and removes a
column from the table body by column name}
\usage{
gt_remove_stat(gts_object, col_name = "stat_0")
}
\arguments{
\item{gts_object}{A data frame, passed by the gtsummary::add_stat function}

\item{col_name}{the column name from the gtsummary object's table_body to remove}
}
\value{
a gtsummary object without the named column
}
\description{
A gtsummary wrapper function that takes a gtsummary object and removes a
column from the table body by column name
}
