% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_survey.R
\name{tabulate_survey}
\alias{tabulate_survey}
\title{Tabulate survey design objects by a categorical and another stratifying variable}
\usage{
tabulate_survey(
  x,
  var,
  strata = NULL,
  pretty = TRUE,
  wide = TRUE,
  digits = 1,
  method = "logit",
  na.rm = FALSE,
  deff = FALSE,
  proptotal = FALSE,
  rowtotals = FALSE,
  coltotals = FALSE
)
}
\arguments{
\item{x}{a survey design object}

\item{var}{the bare name of a categorical variable}

\item{strata}{a variable to stratify the results by}

\item{pretty}{if \code{TRUE}, default, the proportion and CI are merged}

\item{wide}{if \code{TRUE} (default) and strata is defined, then the results are
presented in a wide table with each stratification counts and estimates in
separate columns. If \code{FALSE}, then the data will be presented in a long
format where the counts and estimates are presented in single columns. This
has no effect if strata is not defined.}

\item{digits}{if \code{pretty = FALSE}, this indicates the number of digits used
for proportion and CI}

\item{method}{a method from \code{\link[survey:svyciprop]{survey::svyciprop()}} to calculate the confidence
interval. Defaults to "logit"}

\item{na.rm}{When \code{TRUE}, missing (NA) values present in \code{var} will be removed
from the data set with a warning, causing a change in denominator for the
tabulations.  The default is set to \code{FALSE}, which creates an explicit
missing value called "(Missing)".}

\item{deff}{a logical indicating if the design effect should be reported.
Defaults to "TRUE"}

\item{proptotal}{if \code{TRUE} and \code{strata} is not \code{NULL}, then the totals of the
rows will be reported as proportions of the total data set, otherwise, they
will be proportions within the stratum (default).}

\item{rowtotals}{if \code{TRUE} and \code{strata} is defined, then an extra "Total"
column will be added tabulating all of the rows across strata.}

\item{coltotals}{if \code{TRUE} a new row with totals for each "n" column is
created.}
}
\value{
a long or wide tibble with tabulations n, ci, and deff
}
\description{
This function has been superseeded by \code{\link[=tab_survey]{tab_survey()}} please use that function
instead.
}
\seealso{
\code{\link[epikit:rename_redundant]{epikit::rename_redundant()}}, \code{\link[epikit:rename_redundant]{epikit::augment_redundant()}}
}
\keyword{internal}
