% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_cmh.R
\name{tbl_cmh}
\alias{tbl_cmh}
\title{Stratified Odds Ratios}
\usage{
tbl_cmh(
  data,
  case,
  exposure,
  strata,
  label = NULL,
  estimate_fun = gtsummary::style_ratio,
  overall_label = "Crude"
)
}
\arguments{
\item{data}{a data frame}

\item{case}{case variable}

\item{exposure}{exposure variable}

\item{strata}{stratifying variables}

\item{label}{list of formulas specifying labels for strata}

\item{estimate_fun}{function styling the odds ratios.
Default is \code{gtsummary::style_ratio}}

\item{overall_label}{string indicating the un-stratified ORs.
Default is \code{"Crude"}}
}
\value{
gtsummary table
}
\description{
Stratified Odds Ratios
}
\examples{
library(gtsummary)

tbl_cmh_ex1 <-
  trial \%>\%
  tbl_cmh(case = death, exposure = trt, strata = c(stage, grade))
}
\author{
Daniel D. Sjoberg
}
