% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_univariate_stratified.R
\name{tbl_cmh_single}
\alias{tbl_cmh_single}
\title{internal version of the above function for a single exposure variable}
\usage{
tbl_cmh_single(
  data,
  case,
  exposure,
  strata,
  measure,
  obstime = NULL,
  conf.level = 0.95
)
}
\description{
internal version of the above function for a single exposure variable
}
