% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_univariate_crosstabs.R
\name{add_crosstabs}
\alias{add_crosstabs}
\title{A \{gtsummary\} wrapper function that takes a gtsummary univariate regression
table and adds appropriate cross tabs by exposure and outcome}
\usage{
add_crosstabs(x, wide = FALSE)
}
\arguments{
\item{x}{Object with class \code{tbl_uvregression} from the gtsummary
tbl_uvregression function or \code{tbl_cmh} from the epitabulate tbl_cmh function.}

\item{wide}{TRUE/FALSE to specify whether would like to have the output in
wide format. Results in four columns rather than two, but in a single row.
This is only works for dichotomous variables (yes/no, TRUE/FALSE,
male/female), others will be dropped with a warning message. (Default is FALSE)}
}
\description{
A \{gtsummary\} wrapper function that takes a gtsummary univariate regression
table and adds appropriate cross tabs by exposure and outcome
}
\references{
Inspired by Daniel Sjoberg,
see \href{https://github.com/ddsjoberg/gtsummary}{gtsummary github repo}
}
