% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_wrappers.R
\name{gtsummary_attack_rate}
\alias{gtsummary_attack_rate}
\alias{gtsummary_case_fatality_rate}
\alias{gt_remove_stat}
\alias{add_gt_cfr_stat_label}
\alias{add_gt_cfr_stat_level}
\alias{add_gt_attack_rate_label}
\alias{add_gt_attack_rate_level}
\title{An attack rate wrapper function (using gtsummary and epikit packages)that takes
a gtsummary object and returns a gtsummary object withattack rate (per given
multiple) with 95\% confidence interval}
\usage{
gtsummary_attack_rate(gts_object, population, multiplier)

gtsummary_case_fatality_rate(gts_object, deaths_var)

gt_remove_stat(gt_object, col_name = "stat_0")

add_gt_cfr_stat_label(data, variable, by, deaths_var, ...)

add_gt_cfr_stat_level(data, variable, by, deaths_var, ...)

add_gt_attack_rate_label(
  data,
  variable,
  by = NULL,
  population,
  multiplier,
  drop_population = TRUE,
  ...
)

add_gt_attack_rate_level(
  data,
  variable,
  by = NULL,
  population,
  multiplier,
  ...
)
}
\arguments{
\item{population}{the number of individuals in the population, passed to
\code{epikit::attack_rate}}

\item{multiplier}{The base by which to multiply the output:}

\item{deaths_var}{the name of a logical column in the data that indicates that the case died,
is passed as the first argument to \code{epikit::case_fatality_rate_df}}

\item{gt_object}{A data frame, passed by the gtsummary::add_stat function}

\item{col_name}{the column name from the gtsummary object's table_body to remove}

\item{data}{A data frame, passed by the gtsummary::add_stat function.}

\item{variable}{Name of a variable as the outcome of interest, passed by the
gtsummary::add_stat function (e.g. illness)}

\item{by}{Name of a variable for stratifying, passed by the gtsummary::add_stat
function (e.g. illness).

#'@param population the number of individuals in the population, passed to
\code{epikit::attack_rate}}

\item{...}{additional params that may be passed from gtsummary functions.}

\item{drop_population}{whether or not to include the population column; default TRUE}
}
\value{
a single-row gtsummary object with attack rate results cases,
population, attack rate, and 95\% confidence interval.

a single row gtsummary object with case fatality rate results for
deaths, cases, cfr, and 95\% confidence interval.

a gtsummary object without the named column

a single row gtsummary object with case fatality rate results for
deaths, cases, cfr, and 95\% confidence interval.

a single row or multiple row gtsummary object with with case fatality
rate results for deaths, cases, cfr, and 95\% confidence interval.

a single row gtsummary object with with attack rate results with
cases, attack rate, and 95\% confidence interval.

a single-row gtsummary object with attack rate results cases,
population, attack rate, and 95\% confidence interval.
}
\description{
An attack rate wrapper function (using gtsummary and epikit packages)that takes
a gtsummary object and returns a gtsummary object withattack rate (per given
multiple) with 95\% confidence interval

An case fatality rate wrapper function (using gtsummary and epikit packages)
that takes a gtsummary object and returns a gtsummary object with number
of deaths, case fatality rate, and 95\% confidence interval.

A gtsummary wrapper function that takes a gtsummary object and removes a
column from the table body by column name

A case fatality rate wrapper function to be passed to the gtsummary::add_stat
function, which returns a data frame with a single row to be used with
dichotomous data or overall data.Calls epikit::case_fatality_rate_df.

A case fatality rate wrapper function to be passed to the gtsummary::add_stat function,
which returns a data frame with multiple rows to be used when location is set
to "level" for multi-level categorical data. Calls epikit::case_fatality_rate_df.

An attack rate wrapper function to be passed to the gtsummary::add_stat
function, which returns a data frame with a single row to be used with
dichotomous data or overall data. Calls epikit::attack_rate.

An attack rate wrapper function to be passed to the gtsummary::add_stat function,
which returns a data frame with multiple rows to be used when location is set
to "level" for multi-level categorical data. Calls epikit::attack_rate.
}
